move insert functions to their own file
