+++ backend/main/server/resources/Announcement.py
@@ -6,17 +6,6 @@ from main.server.models import Announcement, AnnouncementSchema
 announcement_schema = AnnouncementSchema()
 announcements_schema = AnnouncementSchema(many=True)
 
-def insertAnnouncement(col, data):
-    if col[0] != "message": return 2
-    if not data or not data[0]: return 2
-    text = data[0]
-    message = Announcement.query.filter_by(message=text).first()
-    if message:
-        return 1
-    message = Announcement(message=text)
-    db.session.add(message)
-    return 0
-
 @app.after_request
 def add_header(response):
     response.headers['Access-Control-Allow-Origin'] = '*'
