+++ backend/main/server/resources/Gallery.py
@@ -6,24 +6,6 @@ from main.server.models import Gallery, GallerySchema
 artwork_schema = GallerySchema()
 gallery_schema = GallerySchema(many=True)
 
-def insertGallery(col, data):
-    artworkLink = artistLink = title = username = None
-    for i in range(0, len(data)):
-        if col[i] == "artworkLink" and data[i]: artworkLink = data[i]
-        elif col[i] == "artistLink" and data[i]: artistLink = data[i]
-        elif col[i] == "title" and data[i]: title = data[i]
-        elif col[i] == "username" and data[i]: username = data[i]
-        else: continue
-    message = Gallery.query.filter_by(artworkLink=artworkLink).first()
-    if message:
-        return 1
-    message = Gallery(artworkLink=artworkLink,
-                          artistLink=artistLink,
-                          username=username,
-                          title=title)
-    db.session.add(message)
-    return 0
-
 @app.after_request
 def add_header(response):
     response.headers['Access-Control-Allow-Origin'] = '*'
