+++ backend/main/server/resources/Game.py
@@ -4,29 +4,9 @@ from flask import request
 from main.server import app, cache, db
 from main.server.models import Games, GameSchema
 
-
 games_schema = GameSchema(many=True)
 game_schema = GameSchema()
 
-def insertGame(col, data):
-    gameLink = gitLink = description = title = thumbnail = None
-    for i in range(0, len(data)):
-        if col[i] == "gameLink" and data[i]: gameLink = data[i]
-        elif col[i] == "gitLink": gitLink = data[i]
-        elif col[i] == "description": description = data[i]
-        elif col[i] == "title" and data[i]: title = data[i]
-        elif col[i] == "thumbnail": thumbnail = data[i]
-        else: continue
-    message = Games.query.filter_by(gameLink=gameLink).first()
-    if message:
-        return 1
-    message = Games(gameLink=gameLink,
-                    gitLink=gitLink,
-                    description=description,
-                    title=title,
-                    thumbnail=thumbnail)
-    db.session.add(message)
- 
 @app.after_request
 def add_header(response):
     response.headers['Access-Control-Allow-Origin'] = '*'
@@ -35,7 +15,6 @@ def add_header(response):
     response.headers[
         'Access-Control-Allow-Headers'] = 'Access-Control-Allow-Headers, Origin,Accept, X-Requested-With, Content-Type, Access-Control-Request-Method, Access-Control-Request-Headers'
     return response
-
 
 class GameListResource(Resource):
     @cache.cached(timeout=100)
